requirements.txt
Authlib
Click==7.0
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python
mysqlclient
peewee
PyJWT
pytz
pymongo==3.10.1
py-jwt-validator
cs50
Flask
Flask-QRcode
Flask-Session
flask-ssl
requests
werkzeug==0.16.1

